change password functionality updated

+++ src/components/NavBar/UserSettings/DeleteAccountPage.jsx
@@ -3,17 +3,18 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { deleteAccount } from '../../../utilities/users-service';
 
-export default function DeleteAccountPage() {
+export default function DeleteAccountPage({setUser}) {
     // Show the confirmation form when delete button is clicked
     const [showConfirmation, setShowConfirmation] = useState(false);
     const navigate = useNavigate();
 
-    const handleDelete = (evt) => {
+    const handleDelete = async (evt) => {
         evt.preventDefault();
         setShowConfirmation(true);
         // allow account to be deleted in this page by calling the deleteAccount API
-        const response = deleteAccount();
+        const response = await deleteAccount();
         if (response.ok) {
+            setUser(null);
             // Redirect to the login page
             navigate('/');
         } else {
@@ -24,7 +25,7 @@ export default function DeleteAccountPage() {
 
     // const handleCancel = async (evt) => {
     //     evt.preventDefault();
-    //     // Redirect to the user settings page
+    //     // Redirect to the user settings page      // onClick={() => evt.preventDefault()}
     //     navigate('/DeleteAccountPage');
     // }
 
@@ -37,13 +38,13 @@ export default function DeleteAccountPage() {
             <h1>Delete User Account</h1>
             <button type="submit" onClick={handleDelete}>Delete</button>
 
-            {showConfirmation && (
-            <form action="" method="DELETE">
-                <h4>Are you sure you want to delete your account?</h4>
-                <button type="submit">Yes - Delete!</button>
-                <button type="button" onClick={handleCancel}>Cancel</button>
-            </form>
-            )}
+                {showConfirmation && (
+                    <form action="" method="DELETE">
+                        <h4>Are you sure you want to delete your account?</h4>
+                        <button type="submit">Yes - Delete!</button>  
+                        <button type="button" onClick={handleCancel}>Cancel</button>
+                    </form>
+                )}
         </div>
     );
 };
